:truck: Restructure to make it more testable
