+++ src/helpers/plainAddPlaceholder/getIndexFromRef.js
@@ -1,0 +1,20 @@
+import SignPdfError from '../../SignPdfError';
+
+/**
+ * @param {object} refTable
+ * @param {string} ref
+ * @returns {number}
+ */
+const getIndexFromRef = (refTable, ref) => {
+    let [index] = ref.split(' ');
+    index = parseInt(index);
+    if (!refTable.offsets.has(index)) {
+        throw new SignPdfError(
+            `Failed to locate object "${ref}".`,
+            SignPdfError.TYPE_PARSE,
+        );
+    }
+    return index;
+};
+
+export default getIndexFromRef;
