+++ src/helpers/plainAddPlaceholder/index.test.js
@@ -1,9 +1,9 @@
 import fs from 'fs';
-import plainAddPlaceholder from './plainAddPlaceholder';
+import plainAddPlaceholder from './index';
 
 describe('plainAddPlaceholder', () => {
     it('adds placeholder to a prepared document', () => {
-        const input = fs.readFileSync(`${__dirname}/../../resources/w3dummy.pdf`);
+        const input = fs.readFileSync(`${__dirname}/../../../resources/w3dummy.pdf`);
         expect(input.indexOf('/ByteRange')).toBe(-1);
         const output = plainAddPlaceholder(input, {reason: 'Because I can'});
         expect(output).toBeInstanceOf(Buffer);
